Fix rdf median values to separate out same locations with different site codes. Continue working on calculation.

+++ rdf_etl/database/migrations/2016_10_07_122347_add-limits-column-to-regional-load-table.php
@@ -13,7 +13,7 @@ class AddLimitsColumnToRegionalLoadTable extends Migration
     public function up()
     {
         Schema::table('regional_load', function ($table) {
-            $table->integer('hourly_limit');
+            $table->decimal('hourly_limit',8,2);
         });
     }
